prepare to test in lambda
