+++ lambda/package.json
@@ -23,8 +23,7 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "openai": "^3.2.1",
-    "pdf2json": "2.0.1",
-    "pdfjs-dist": "^3.4.120"
+    "pdf2json": "2.0.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.17",
