+++ lambda/src/index.ts
@@ -1,5 +1,5 @@
-import { generateCoverLetter } from "./services/openai";
-import { getContent } from "./services/pdf";
+// import { generateCoverLetter } from "./services/openai";
+import { parsePdf } from "./services/parsePdf";
 
 // TODO: look into this https://haydnjmorris.medium.com/aws-lambda-node-modules-no-docker-required-76443a0f6c4e
 // I am having an issue with doing npm install locally on windows, then trying to use those node_modules in a lambda
@@ -7,16 +7,21 @@ exports.handler = async (event: { body: string; }, _: any) => {
     console.log('I AM IN THE INDEX FILE!!!')
     let response;
     try {
-        const body = JSON.parse(event.body);
-        const { jobDescription } = body;
-        console.log('body', body);
-        const prompt = body.prompt || await getContent(body.resumeUrl);
-        console.info("prompt:", prompt)
-        const coverLetter = await generateCoverLetter(jobDescription, prompt);
-        if (!coverLetter) {
-            throw new Error("ChatGPT could not create a cover letter from the provided information");
-        }
+        console.log('event', event);
+        // const body = JSON.parse(event.body);
+        // const { jobDescription } = body;
+        // console.log('body', body);
+        // const prompt = body.prompt;
+        // console.info("prompt:", prompt)
+        // test pdf parsing in lambda
+        const resumeText = await parsePdf();
+        console.log('Resume text in main handler: ', resumeText);
+        // const coverLetter = await generateCoverLetter(jobDescription, prompt);
+        // if (!coverLetter) {
+        //     throw new Error("ChatGPT could not create a cover letter from the provided information");
+        // }
         
+        const coverLetter = 'This is a cover letter'
         response = {
             statusCode: 200,
             body: JSON.stringify({
